fix withdraw sum to float

+++ internal/models/user.go
@@ -31,8 +31,8 @@ type UserBalanceResponse struct {
 }
 
 type WithdrawRequest struct {
-	Order string `json:"order"`
-	Sum   int    `json:"sum"`
+	Order string  `json:"order"`
+	Sum   float64 `json:"sum"`
 }
 
 type WithdrawalsResponse struct {
